refactor: :truck: move resource
